requirements.txt
streamlit
langchain==0.1.20
langchain-community==0.0.38
sumy
transformers
lexrank
datasets
rouge
rouge_score
nltk==3.8.1
langchain-huggingface
transformers 
torch
python-docx

